Fix reward items number mapping

+++ domain/src/main/java/onlydust/com/marketplace/api/domain/view/RewardItemView.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @Builder
 @Data
 public class RewardItemView {
-    Integer number;
+    Long number;
     String id;
     String title;
     String githubUrl;
